+++ simplq/src/components/pages/Home/MyQueues.jsx
@@ -2,7 +2,8 @@ import React, { useEffect, useCallback } from 'react';
 import DeleteIcon from '@material-ui/icons/Delete';
 import IconButton from '@material-ui/core/IconButton';
 import { useHistory } from 'react-router';
-import { useFetchQueues, useDeleteQueue, selectQueues } from 'store/queues';
+import { useGetUserQueues, useDeleteQueue } from 'store/asyncActions';
+import { selectQueues } from 'store/queues';
 import { useDispatch, useSelector } from 'react-redux';
 import styles from './home.module.scss';
 
@@ -10,12 +11,12 @@ export default () => {
   const history = useHistory();
   const dispatch = useDispatch();
   const queues = useSelector(selectQueues);
-  const fetchQueues = useCallback(useFetchQueues(), []);
+  const getUserQueues = useCallback(useGetUserQueues(), []);
   const deleteQueue = useDeleteQueue();
 
   useEffect(() => {
-    dispatch(fetchQueues());
-  }, [dispatch, fetchQueues]);
+    dispatch(getUserQueues());
+  }, [dispatch, getUserQueues]);
 
   const handleDelete = (e, queue) => {
     // Don't trigger parent's onClick
